Per #38, suffix_list_file -> suffix_list_url.
Rename keyword argument for TLDExtract from `suffix_list_file` to
`suffix_list_url`. Update docstring accordingly. This makes it so
that it is compatible with the `fetch` keyword argument.

(Of course this means that to use your own, local `suffix_list_file`,
you should specify the path to that file using `file://` protocol.
Addition to the README, re: this usage, is forthcoming in another
commit.)

Also add `debug` keyword argument, to allow users to disable
the diff-from-snapshot logging (that's a lot of logging to happen by
default, and it's not always relevant to people who are using `tldextract`
as a libary in a larger system). (Who haven't taken the time to set up
a logger for said library ;) )