Support IPv6 addresses
IPv6 addresses within URLs must be wrapped in `[]` so we can partition on `]` once the netloc has been isolated.
If the IPv6 address is not part of a URL (or is not wrapped in brackets) it cannot have a port associated with it, so we can check for a single `:` and partition on that. If it has more than one colon it, it is likely to be an IPv6 address (unless it is just a malformed hostname string). I have added extra checking for IPv6 addresses in the section where IPv4 addresses are tested, however this is not strictly necessary since they would be placed in the domain location of an `ExtractResult` anyway.
I have also reordered the way partitioning of a URL is handled since it makes more sense to remove elements from right to left (i.e. fragment first, then query and path) until we are left with the netloc.