Revert "Pylint compat with Python 2.6"
Jumping through hoops for an old version of Python isn't worth it. Pylint 1.4 is superior to Pylint 1.3. It's the same code for all versions. If we have to skip Python 2.6 lint checks, so be it.

This mostly reverts commit 888291372d5504ae906610aac9303c733db7092f. Keep the Pylint subprocess stderr checking.