Require docstrings in tests
Sometimes, a Python function name doesn't tell you all the arcane stuff going on inside a test case. Require test case and test helper descriptions going forward. This commit doesn't perfectly describe all 10+ years of tests in this project, some of that context is lost to the ages, but it's a start.

* Combine and rename some duplicate test cases